Adjust resize event to be based on resolution change
The resize event of SpriteScreen first before the initial window size is
determined. Therefore the content of GameService.Graphics.Resolution
contains a wrong value. Because there is no resize event for the window
itself and also no ResolutionChanged event - this implementation
simulates such an event to update the clock bounds only if the
Resolution has changed not the SpriteScreen size.